pyproject.toml
[project]
name = "spde-gp-kernel"
version = "0.1.0"
description = "Sparse variational GP state-space models with PyTorch"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "Taylor Mason"}]
dependencies = [
    "torch>=2.2.0,<2.3.0",
    "pyro-ppl>=1.9.1,<1.10",
    "numpy>=1.26",
    "pyyaml>=6.0",
    "typer>=0.12",
    "rich>=13.7"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "black>=24.4",
    "ruff>=0.5",
    "mypy>=1.10"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
